Percentage values fixes

+++ HeroesMatchTracker.Core/Models/StatisticsModels/StatsHeroesTalents.cs
@@ -11,6 +11,6 @@ namespace HeroesMatchTracker.Core.Models.StatisticsModels
         public int Wins { get; set; }
         public int Losses { get; set; }
         public int Total { get; set; }
-        public int Winrate { get; set; }
+        public double Winrate { get; set; }
     }
 }
